+++ RELEASES.md
@@ -15,6 +15,7 @@ None
 
 ### Fixes
 - Fixed trailing stop market fill behavior when top-level exhausted to align with market orders (#2540), thanks for reporting @stastnypremysl
+- Fixed terminating backtest on `AccountError` when streaming, the exception needed to be reraised to interrupt the streaming of chunks (#2546), thanks for reporting @stastnypremysl
 
 ### Documentation Updates
 None